shifts TFNadmin editeor back into darkpack one
